+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/smallInteger/AddNode.java
@@ -1,4 +1,4 @@
-package org.enso.interpreter.node.expression.builtin.number.int64;
+package org.enso.interpreter.node.expression.builtin.number.smallInteger;
 
 import com.oracle.truffle.api.dsl.Fallback;
 import com.oracle.truffle.api.dsl.Specialization;
@@ -9,38 +9,38 @@ import org.enso.interpreter.node.expression.builtin.number.utils.ToEnsoNumberNod
 import org.enso.interpreter.runtime.error.TypeError;
 import org.enso.interpreter.runtime.number.EnsoBigInteger;
 
-@BuiltinMethod(type = "Small_Integer", name = "-", description = "Subtraction of numbers.")
-public abstract class SubtractNode extends Node {
+@BuiltinMethod(type = "Small_Integer", name = "+", description = "Addition of numbers.")
+public abstract class AddNode extends Node {
   private @Child ToEnsoNumberNode toEnsoNumberNode = ToEnsoNumberNode.build();
 
   abstract Object execute(long _this, Object that);
 
-  static SubtractNode build() {
-    return SubtractNodeGen.create();
+  static AddNode build() {
+    return AddNodeGen.create();
   }
 
   @Specialization(rewriteOn = ArithmeticException.class)
   long doLong(long _this, long that) {
-    return Math.subtractExact(_this, that);
+    return Math.addExact(_this, that);
   }
 
   @Specialization
   Object doOverflow(long _this, long that) {
-    return toEnsoNumberNode.execute(BigIntegerOps.subtract(_this, that));
+    return toEnsoNumberNode.execute(BigIntegerOps.add(_this, that));
   }
 
   @Specialization
   double doDouble(long _this, double that) {
-    return _this - that;
+    return _this + that;
   }
 
   @Specialization
   Object doBigInteger(long _this, EnsoBigInteger that) {
-    return toEnsoNumberNode.execute(BigIntegerOps.subtract(_this, that.getValue()));
+    return toEnsoNumberNode.execute(BigIntegerOps.add(that.getValue(), _this));
   }
 
   @Fallback
   Object doOther(long _this, Object that) {
-    throw new TypeError("Unexpected type provided for argument `that` in Integer.-", this);
+    throw new TypeError("Unexpected type provided for argument `that` in Integer.+", this);
   }
 }
