+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/smallInteger/DivNode.java
@@ -1,4 +1,4 @@
-package org.enso.interpreter.node.expression.builtin.number.int64;
+package org.enso.interpreter.node.expression.builtin.number.smallInteger;
 
 import com.oracle.truffle.api.dsl.Fallback;
 import com.oracle.truffle.api.dsl.Specialization;
@@ -7,26 +7,26 @@ import org.enso.interpreter.dsl.BuiltinMethod;
 import org.enso.interpreter.runtime.error.TypeError;
 import org.enso.interpreter.runtime.number.EnsoBigInteger;
 
-@BuiltinMethod(type = "Small_Integer", name = "%", description = "Modulo division of numbers.")
-public abstract class ModNode extends Node {
+@BuiltinMethod(type = "Small_Integer", name = "div", description = "Division of numbers.")
+public abstract class DivNode extends Node {
   abstract Object execute(long _this, Object that);
 
-  static ModNode build() {
-    return ModNodeGen.create();
+  static DivNode build() {
+    return DivNodeGen.create();
   }
 
   @Specialization
   long doLong(long _this, long that) {
-    return _this % that;
+    return _this / that;
   }
 
   @Specialization
   long doBigInteger(long _this, EnsoBigInteger that) {
-    return _this;
+    return 0L;
   }
 
   @Fallback
   Object doOther(long _this, Object that) {
-    throw new TypeError("Unexpected type provided for argument `that` in Integer.%", this);
+    throw new TypeError("Unexpected type provided for argument `that` in Integer.div", this);
   }
 }
