+++ engine/runtime/src/main/java/org/enso/interpreter/node/expression/builtin/number/smallInteger/MultiplyNode.java
@@ -1,4 +1,4 @@
-package org.enso.interpreter.node.expression.builtin.number.int64;
+package org.enso.interpreter.node.expression.builtin.number.smallInteger;
 
 import com.oracle.truffle.api.dsl.Fallback;
 import com.oracle.truffle.api.dsl.Specialization;
@@ -9,38 +9,38 @@ import org.enso.interpreter.node.expression.builtin.number.utils.ToEnsoNumberNod
 import org.enso.interpreter.runtime.error.TypeError;
 import org.enso.interpreter.runtime.number.EnsoBigInteger;
 
-@BuiltinMethod(type = "Small_Integer", name = "+", description = "Addition of numbers.")
-public abstract class AddNode extends Node {
+@BuiltinMethod(type = "Small_Integer", name = "*", description = "Multiplication of numbers.")
+public abstract class MultiplyNode extends Node {
   private @Child ToEnsoNumberNode toEnsoNumberNode = ToEnsoNumberNode.build();
 
   abstract Object execute(long _this, Object that);
 
-  static AddNode build() {
-    return AddNodeGen.create();
+  static MultiplyNode build() {
+    return MultiplyNodeGen.create();
   }
 
   @Specialization(rewriteOn = ArithmeticException.class)
   long doLong(long _this, long that) {
-    return Math.addExact(_this, that);
+    return Math.multiplyExact(_this, that);
   }
 
   @Specialization
   Object doOverflow(long _this, long that) {
-    return toEnsoNumberNode.execute(BigIntegerOps.add(_this, that));
+    return toEnsoNumberNode.execute(BigIntegerOps.multiply(_this, that));
   }
 
   @Specialization
   double doDouble(long _this, double that) {
-    return _this + that;
+    return ((double) _this) * that;
   }
 
   @Specialization
   Object doBigInteger(long _this, EnsoBigInteger that) {
-    return toEnsoNumberNode.execute(BigIntegerOps.add(that.getValue(), _this));
+    return toEnsoNumberNode.execute(BigIntegerOps.multiply(that.getValue(), _this));
   }
 
   @Fallback
   Object doOther(long _this, Object that) {
-    throw new TypeError("Unexpected type provided for argument `that` in Integer.+", this);
+    throw new TypeError("Unexpected type provided for argument `that` in Integer.*", this);
   }
 }
